requirements.txt
# Questi sono di base per tutti
Django>=1.8
psycopg2>=2.5.2
djangorestframework>=3.2.3
django-rest-swagger>=0.3.4
django-filter>=0.9.2
django-oauth-toolkit>=0.9.0
openpyxl>=2.1.3

# WARN!!! See README.md file for problems here!!! 
lxml>=3.4.2
